+++ packages/nimble-components/src/table/index.ts
@@ -28,7 +28,7 @@ import {
     ExpandedState as TanStackExpandedState,
     OnChangeFn as TanStackOnChangeFn
 } from '@tanstack/table-core';
-import { keyShift } from '@microsoft/fast-web-utilities';
+import { keyEnter, keyShift } from '@microsoft/fast-web-utilities';
 import { TableColumn } from '../table-column/base';
 import { TableValidator } from './models/table-validator';
 import { styles } from './styles';
@@ -54,7 +54,8 @@ import { Virtualizer } from './models/virtualizer';
 import { getTanStackSortingFunction } from './models/sort-operations';
 import { TableLayoutManager } from './models/table-layout-manager';
 import { TableUpdateTracker } from './models/table-update-tracker';
-import type { TableRow } from './components/row';
+import { TableRow } from './components/row';
+import type { TableGroupRow } from './components/group-row';
 import { ColumnInternals } from '../table-column/base/models/column-internals';
 import { InteractiveSelectionManager } from './models/interactive-selection-manager';
 import { DataHierarchyManager } from './models/data-hierarchy-manager';
@@ -62,6 +63,8 @@ import { ExpansionManager } from './models/expansion-manager';
 import { waitUntilCustomElementsDefinedAsync } from '../utilities/wait-until-custom-elements-defined-async';
 import { ColumnValidator } from '../table-column/base/models/column-validator';
 import { uniquifySlotNameForColumnId } from './models/utilities';
+import { KeyboardNavigationManager } from './models/keyboard-navigation-manager';
+import { TableCellView } from '../table-column/base/cell-view';
 
 declare global {
     interface HTMLElementTagNameMap {
@@ -106,7 +109,7 @@ export class Table<
      * @internal
      */
     @observable
-    public readonly rowElements: TableRow[] = [];
+    public readonly rowElements: (TableRow | TableGroupRow)[] = [];
 
     /**
      * @internal
@@ -173,6 +176,12 @@ export class Table<
      * @internal
      */
     @observable
+    public readonly collapseAllButton?: HTMLElement;
+
+    /**
+     * @internal
+     */
+    @observable
     public showCollapseAll = false;
 
     /**
@@ -199,6 +208,11 @@ export class Table<
      * @internal
      */
     public readonly layoutManager: TableLayoutManager<TData>;
+
+    /**
+     * @internal
+     */
+    public readonly keyboardNavigationManager: KeyboardNavigationManager<TData>;
 
     /**
      * @internal
@@ -275,6 +289,10 @@ export class Table<
         };
         this.table = tanStackCreateTable(this.options);
         this.virtualizer = new Virtualizer(this, this.table);
+        this.keyboardNavigationManager = new KeyboardNavigationManager(
+            this,
+            this.virtualizer
+        );
         this.layoutManager = new TableLayoutManager(this);
         this.layoutManagerNotifier = Observable.getNotifier(this.layoutManager);
         this.layoutManagerNotifier.subscribe(this, 'isColumnBeingSized');
@@ -329,6 +347,7 @@ export class Table<
         this.viewport.addEventListener('scroll', this.onViewPortScroll, {
             passive: true
         });
+        this.keyboardNavigationManager.connect();
         document.addEventListener('keydown', this.onKeyDown);
         document.addEventListener('keyup', this.onKeyUp);
     }
@@ -336,6 +355,7 @@ export class Table<
     public override disconnectedCallback(): void {
         super.disconnectedCallback();
         this.virtualizer.disconnect();
+        this.keyboardNavigationManager.disconnect();
         this.viewport.removeEventListener('scroll', this.onViewPortScroll);
         document.removeEventListener('keydown', this.onKeyDown);
         document.removeEventListener('keyup', this.onKeyUp);
@@ -406,6 +426,16 @@ export class Table<
         }
 
         return true;
+    }
+
+    /** @internal */
+    public onRowFocusIn(event: FocusEvent): void {
+        this.keyboardNavigationManager.onRowFocusIn(event);
+    }
+
+    /** @internal */
+    public onRowBlur(event: FocusEvent): void {
+        this.keyboardNavigationManager.onRowBlur(event);
     }
 
     /** @internal */
@@ -562,6 +592,19 @@ export class Table<
     /**
      * @internal
      */
+    public onHeaderKeyDown(column: TableColumn, event: KeyboardEvent): boolean {
+        const allowMultiSort = event.shiftKey;
+        if (event.key === keyEnter) {
+            this.toggleColumnSort(column, allowMultiSort);
+        }
+        // Return true so that we don't prevent default behavior. Without this, Tab navigation
+        // gets stuck on the column headers.
+        return true;
+    }
+
+    /**
+     * @internal
+     */
     public update(): void {
         this.validate();
         if (this.tableUpdateTracker.requiresTanStackUpdate) {
@@ -577,6 +620,10 @@ export class Table<
             this.visibleColumns = this.columns.filter(
                 column => !column.columnHidden
             );
+        }
+
+        if (this.tableUpdateTracker.requiresKeyboardFocusReset) {
+            this.keyboardNavigationManager.resetFocusState();
         }
     }
 
@@ -635,6 +682,37 @@ export class Table<
             };
         }
         return tanStackUpdates;
+    }
+
+    /** @internal */
+    public handleFocusedCellRecycling(): void {
+        const hadActiveRowOrCellFocus = this.keyboardNavigationManager.hasActiveRowOrCellFocus;
+
+        let tableFocusedElement = this.shadowRoot!.activeElement;
+        while (
+            tableFocusedElement !== null
+            && !(tableFocusedElement instanceof TableCellView)
+        ) {
+            if (tableFocusedElement.shadowRoot) {
+                tableFocusedElement = tableFocusedElement.shadowRoot.activeElement;
+            } else {
+                break;
+            }
+        }
+        if (tableFocusedElement instanceof TableCellView) {
+            tableFocusedElement.focusedRecycleCallback();
+        }
+        if (this.openActionMenuRecordId !== undefined) {
+            const activeRow = this.rowElements.find(
+                row => row instanceof TableRow
+                    && row.recordId === this.openActionMenuRecordId
+            ) as TableRow | undefined;
+            activeRow?.closeOpenActionMenus();
+        }
+
+        this.keyboardNavigationManager.handleFocusedCellRecycling(
+            hadActiveRowOrCellFocus
+        );
     }
 
     protected selectionModeChanged(
@@ -710,6 +788,7 @@ export class Table<
     private async handleRowActionMenuToggleEvent(
         event: CustomEvent<TableActionMenuToggleEventDetail>
     ): Promise<void> {
+        this.keyboardNavigationManager.onRowActionMenuToggle(event);
         const detail = await this.getActionMenuToggleEventDetail(event);
         this.$emit('action-menu-toggle', detail);
         if (!event.detail.newState) {
@@ -1003,6 +1082,7 @@ export class Table<
                 isParentRow: isParent,
                 immediateChildCount: row.subRows.length,
                 groupColumn: this.getGroupRowColumn(row),
+                resolvedRowIndex: row.index,
                 isLoadingChildren: this.expansionManager.isLoadingChildren(
                     row.id
                 ),