+++ packages/nimble-components/src/table/types.ts
@@ -1,5 +1,8 @@
+import type { Checkbox } from '../checkbox';
+import type { MenuButton } from '../menu-button';
 import type { TableColumn } from '../table-column/base';
 import type { ValidityObject } from '../utilities/models/validator';
+import type { TableCell } from './components/cell';
 
 /**
  * TableFieldName describes the type associated with keys within
@@ -212,14 +215,42 @@ export interface TableRowState<TData extends TableRecord = TableRecord> {
     isParentRow: boolean;
     isLoadingChildren: boolean;
     slots: SlotMetadata[];
+    resolvedRowIndex?: number;
 }
+/**
+ * Table keyboard focus types
+ */
+export const TableFocusType = {
+    none: 'none',
+    columnHeader: 'columnHeader',
+    headerActions: 'headerActions',
+    row: 'row',
+    rowSelectionCheckbox: 'rowSelectionCheckbox',
+    cell: 'cell',
+    cellActionMenu: 'cellActionMenu',
+    cellContent: 'cellContent'
+} as const;
+export type TableFocusType =
+    (typeof TableFocusType)[keyof typeof TableFocusType];
 
 /**
  * @internal
+ * Focusable elements of a table row
  */
-export interface SlotMetadata {
-    slot: string;
-    name: string;
+export interface TableRowFocusableElements {
+    selectionCheckbox?: Checkbox;
+    cells: {
+        cell: TableCell,
+        actionMenuButton?: MenuButton
+    }[];
+}
+
+/**
+ * Focusable elements of a table's header
+ */
+export interface TableHeaderFocusableElements {
+    headerActions: HTMLElement[];
+    columnHeaders: HTMLElement[];
 }
 
 /**
@@ -236,4 +267,12 @@ export interface RowSlotRequestEventDetail {
     columnInternalId: string;
     recordId: string;
     slots: SlotMetadata[];
+}
+
+/**
+ * @internal
+ */
+export interface SlotMetadata {
+    slot: string;
+    name: string;
 }